requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.1
black==25.1.0
boto3==1.37.7
botocore==1.37.7
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
django-ninja==1.3.0
django-storages==1.14.5
flake8==7.1.1
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pybars3==0.9.7
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
PyMeta3==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
